+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.3] - 2026-04-13
+
+### Added
+
+- Table pretty-printing (column alignment) in bat render mode.
+
+### Fixed
+
+- Fix table column misalignment with emoji/wide Unicode characters.
+
 ## [0.9.2] - 2026-04-13
 
 ### Fixed